+++ Iov42sdk/Support/TransferAuthorisations.cs
@@ -2,13 +2,13 @@
 
 namespace Iov42sdk.Support
 {
-    public class TradeAuthorisations
+    public class TransferAuthorisations
     {
         public string RequestId { get; }
         public string BodyText { get; }
         public Authorisation[] Authorisations { get; }
 
-        public TradeAuthorisations(string requestId, string bodyText, params Authorisation[] authorisations)
+        public TransferAuthorisations(string requestId, string bodyText, params Authorisation[] authorisations)
         {
             RequestId = requestId;
             BodyText = bodyText;